% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifts_gender.R
\docType{data}
\name{gifts_gender}
\alias{gifts_gender}
\title{Valentine's Day Gift Spending by Gender}
\format{
A data frame with 2 rows and 9 variables.
\describe{
\item{Gender}{Gender only including Men or Women}
\item{SpendingCelebrating}{Percent spending money on or celebrating Valentines Day}
\item{Candy}{Average percent spending on candy}
\item{Flowers}{Average percent spending on flowers}
\item{Jewelry}{Average percent spending on jewelry}
\item{GreetingCards}{Average percent spending on greeting cards}
\item{EveningOut}{Average percent spending on an evening out}
\item{Clothing}{Average percent spending on clothing}
\item{GiftCards}{Average percent spending on gift cards}
}
}
\source{
\url{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2024/2024-02-13/gifts_gender.csv}
}
\usage{
gifts_gender
}
\description{
A dataset that contains spending on various gift categories during Valentine's Day, breakdown by gender.
}
\keyword{datasets}
