% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_gift_trends_percent_buying.R
\docType{data}
\name{historical_gift_trends_percent_buying}
\alias{historical_gift_trends_percent_buying}
\title{Historical Gift Trends - Percent Buying}
\format{
A data frame with 13 rows and 8 variables.
\describe{
\item{Year}{Year of the recorded spending data}
\item{Candy}{Average percent spending on candy}
\item{Flowers}{Average percent spending on flowers}
\item{Jewelry}{Average percent spending on jewelryy}
\item{Greeting cards}{Average percent spending on greeting cards}
\item{An evening out}{Average percent spending on an evening out}
\item{Clothing}{Average percent spending on clothing}
\item{Gift cards}{Average percent spending on gift cardss}
}
}
\source{
\url{https://www.kaggle.com/datasets/infinator/happy-valentines-day-2022?select=historical_gift_trends_percent_buying.csv}
}
\usage{
historical_gift_trends_percent_buying
}
\description{
A dataset that contains the percentage of people buying different types of gifts for Valentine's Day from 2010 to 2022.
}
\keyword{datasets}
