% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch shiny app}
\usage{
launch_app()
}
\value{
A detailed analysis of Valentine's Day spending trends presented with interactive plots in a dashboard.
}
\description{
Launch shiny app
}
